requirements-dev.txt
PyYAML==3.11
Pygments==2.0.2
argparse==1.4.0
aspy.yaml==0.2.1
blessings==1.6
bpython==0.14.2
cached-property==1.3.0
curtsies==0.1.23
docker-compose==1.5.2
docker-py==1.6.0
dockerpty==0.3.4
docopt==0.6.2
greenlet==0.4.9
jsonschema==2.5.1
nodeenv==0.13.6
ordereddict==1.1
pre-commit==0.7.5
requests==2.7.0
six==1.10.0
texttable==0.8.4
virtualenv==13.1.2
websocket-client==0.35.0
